+++ kelvin-api/ucsschool/kelvin/routers/v1/workgroup.py
@@ -102,14 +102,17 @@ class WorkGroupPatchDocument(KelvinBaseModel):
 
     class Config(UcsSchoolBaseModel.Config):
         lib_class = WorkGroup
+        non_nullable_fields: tuple[str, ...] = ("name", "users")
 
     @validator("name")
-    def check_name(cls, value: str) -> str:
+    def check_name(cls, value: str | None) -> str:
         """
         At this point we know `school` is valid, but
         we don't have it in the values. Thus we use
         the dummy school name DEMOSCHOOL.
         """
+        if value is None:
+            raise ValueError("Null value in property.")
         workgroup_name = f"DEMOSCHOOL-{value}"
         cls.Config.lib_class.name.validate(workgroup_name)
         return value